Convert node handler to sidecar impl
Fixes #31
