+++ function-invokers/node-function-invoker/src/main/docker/.dockerignore
@@ -1,0 +1,1 @@
+image/node_modules
