+++ function-invokers/node-function-invoker/src/main/docker/Dockerfile
@@ -1,0 +1,5 @@
+FROM node:8.6
+ADD image/* /
+RUN npm install
+EXPOSE 8080
+CMD ["node", "server.js"]
