+++ function-invokers/node-function-invoker/src/main/docker/image/package.json
