+++ samples/node/square.js
