+++ samples/square-function.yaml
@@ -3,11 +3,10 @@ kind: Function
 metadata:
   name: square
 spec:
+  image: sk8s/node-function-invoker:latest
+  protocol: http
   input: numbers
   output: greetings
-  handlerRef:
-    name: node
-  params:
-  - name: uri
-    value: /functions/node/hello.js
-
+  env:
+  - name: FUNCTION_URI
+    value: /resources/node/square.js
