+++ teardown
@@ -2,15 +2,18 @@
 
 kubectl delete functions --all
 kubectl delete topics --all
+
 kubectl delete all -l app=topic-controller
 kubectl delete all -l app=topic-gateway
 kubectl delete all -l app=event-dispatcher
-kubectl delete jobs -l function=echo
+
+kubectl delete all -l function=echo
 kubectl delete all -l function=greeter
 kubectl delete all -l function=square
-kubectl delete all -l handler=node
+
 kubectl delete crd/functions.extensions.sk8s.io
 kubectl delete crd/topics.extensions.sk8s.io
+
 kubectl delete all -l app=kafka
 kubectl delete all -l app=zipkin
 kubectl delete all -l app=metrics
